// migrations/snapshots/883ec51ee70e2541928addefd6febafa9ddb62125b7cc9176dc8c4d770330acd/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly BlogCategory: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly BlogComment: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly body: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly postId: CodecTypes['pg/uuid@1']['output'];
      readonly authorId: CodecTypes['pg/uuid@1']['output'];
      readonly parentId: CodecTypes['pg/uuid@1']['output'] | null;
    };
    readonly BlogPost: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly excerpt: CodecTypes['pg/text@1']['output'] | null;
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly coverImage: CodecTypes['pg/text@1']['output'] | null;
      readonly status: 'DRAFT' | 'PUBLISHED' | 'ARCHIVED';
      readonly publishedAt: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly readingTime: CodecTypes['pg/int4@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly authorId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly BlogPostCategory: {
      readonly postId: CodecTypes['pg/uuid@1']['output'];
      readonly categoryId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly BlogPostLike: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly postId: CodecTypes['pg/uuid@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly EmailVerification: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly usedAt: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly PasswordReset: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly usedAt: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly RefreshToken: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly Session: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly ipAddress: CodecTypes['pg/text@1']['output'] | null;
      readonly userAgent: CodecTypes['pg/text@1']['output'] | null;
      readonly expiresAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'] | null;
      readonly passwordHash: CodecTypes['pg/text@1']['output'];
      readonly fullName: CodecTypes['pg/text@1']['output'] | null;
      readonly avatarUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly role: 'STUDENT' | 'INSTRUCTOR' | 'ADMIN';
      readonly isVerified: CodecTypes['pg/bool@1']['output'];
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly lastActiveAt: CodecTypes['pg/timestamptz-string@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly BlogCategory: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly BlogComment: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly body: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly postId: CodecTypes['pg/uuid@1']['input'];
      readonly authorId: CodecTypes['pg/uuid@1']['input'];
      readonly parentId: CodecTypes['pg/uuid@1']['input'] | null;
    };
    readonly BlogPost: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly excerpt: CodecTypes['pg/text@1']['input'] | null;
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly coverImage: CodecTypes['pg/text@1']['input'] | null;
      readonly status: 'DRAFT' | 'PUBLISHED' | 'ARCHIVED';
      readonly publishedAt: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly readingTime: CodecTypes['pg/int4@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly authorId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly BlogPostCategory: {
      readonly postId: CodecTypes['pg/uuid@1']['input'];
      readonly categoryId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly BlogPostLike: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly postId: CodecTypes['pg/uuid@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly EmailVerification: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly usedAt: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly PasswordReset: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly usedAt: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly RefreshToken: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly Session: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly ipAddress: CodecTypes['pg/text@1']['input'] | null;
      readonly userAgent: CodecTypes['pg/text@1']['input'] | null;
      readonly expiresAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'] | null;
      readonly passwordHash: CodecTypes['pg/text@1']['input'];
      readonly fullName: CodecTypes['pg/text@1']['input'] | null;
      readonly avatarUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly role: 'STUDENT' | 'INSTRUCTOR' | 'ADMIN';
      readonly isVerified: CodecTypes['pg/bool@1']['input'];
      readonly isActive: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly lastActiveAt: CodecTypes['pg/timestamptz-string@1']['input'] | null;
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;